--- src/app/(private)/onboarding/family/[mode]/page.tsx
@@ -1,61 +1,0 @@
-"use client";
-
-import Link from "next/link";
-import { FixedInput } from "@/components";
-import { useValidateParams } from "@/hooks/useValidateParams";
-
-const variablesByMode = {
-  create: {
-    guide: "우리 가족 감나무만의 이름을 지어주세요",
-    placeholder: "화목한 김가네 주렁주렁",
-    otherLink: {
-      href: "/onboarding/join/code",
-      text: "이미 감나무 코드를 받으셨나요?",
-    },
-  },
-  code: {
-    guide: "우리 가족의 고유 코드를 입력해 주세요",
-    placeholder: "071409",
-    otherLink: {
-      href: "/onboarding/join/create",
-      text: "아직 개설된 감나무가 없나요?",
-    },
-  },
-} as const;
-
-export default function Page() {
-  const { mode } = useValidateParams<{ mode: "create" | "code" }>({
-    validator: ({ mode }) => ["create", "code"].includes(mode),
-    redirect: "/onboarding/family",
-  });
-
-  return (
-    <div className="flex flex-col justify-center h-full gap-16 px-8">
-      <div>
-        <div className="text-orange-primary font-bold px-2 mb-2 text-center break-keep">
-          {variablesByMode[mode].guide}
-        </div>
-        {mode === "code" ? (
-          <input
-            placeholder={variablesByMode[mode].placeholder}
-            className="styled-input--lg text-center"
-          />
-        ) : (
-          <FixedInput
-            placeholder={variablesByMode[mode].placeholder}
-            suffix="감나무"
-          />
-        )}
-      </div>
-
-      <div className="flex-col-center gap-4">
-        <button className="styled-btn--orange w-full">확인</button>
-        <Link
-          className="text-sm underline text-gray-primary"
-          href={variablesByMode[mode].otherLink.href}>
-          {variablesByMode[mode].otherLink.text}
-        </Link>
-      </div>
-    </div>
-  );
-}
