+++ src/app/(private)/onboarding/family/page.tsx
@@ -1,6 +1,7 @@
 import Link from "next/link";
 import { FaArrowCircleRight } from "react-icons/fa";
 import { twMerge } from "tailwind-merge";
+import { routePaths } from "@/constants/route";
 
 export default function Page() {
   const buttonClassName =
@@ -10,7 +11,7 @@ export default function Page() {
       <div className="flex-col-center gap-2 w-full">
         <div className="text-gray-dark">아직 개설된 감나무가 없나요?</div>
         <Link
-          href="/onboarding/familiy/create"
+          href={routePaths.private.onboardingFamilyCreate}
           className={twMerge(buttonClassName, "bg-green-primary")}>
           우리 가족 감나무 만들기
           <FaArrowCircleRight className="text-2xl" />
@@ -19,7 +20,7 @@ export default function Page() {
       <div className="flex-col-center gap-2 w-full">
         <div className="text-gray-dark">이미 감나무 코드를 받으셨나요?</div>
         <Link
-          href="/onboarding/family/code"
+          href={routePaths.private.onboardingFamilyJoin}
           className={twMerge(buttonClassName, "bg-orange-primary")}>
           코드 입력하고 참여하기
           <FaArrowCircleRight className="text-2xl" />
